+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Fixed `Content.expand_tabs` https://github.com/Textualize/textual/pull/6038
 - Fixed return value for `Pilot.double_click` and `Pilot.triple_click` https://github.com/Textualize/textual/pull/6035
 - Fixed sizing issue with `Pretty` widget https://github.com/Textualize/textual/pull/6040 https://github.com/Textualize/textual/pull/6041
+- Fixed garbled inline app output when `inline_no_clear=True` https://github.com/Textualize/textual/pull/6080
 
 ### Added